Extend the callback types of PINS.
Add some CPP logic to support PINS callbacks of arbitrary
(but fixed) prototype, while keeping some type checking
capability.

The goal is to be able to capture more information in PINS.
Typically, AMT profilers start to gather information like
parent/child relationship. With the current PINS API, it's
hard to impossible to expose that information to a PINS module,
because the only parameters passed to the module are the
execution stream, the user-defined callback data, and a task.

With these changes (which required to change the order of
parameters), we can extend PINS as much as needed to support
an arbitrary callback prototype. This commit shows how
callbacks that never pass a task parameter are supported
by removing that parameter. The next commit will show how
to replace the parsec_prof_grapher interface with a PINS
module using this capability.