Improve efficiency of stream management in GPU.
When a GPU kernel had all its data already on the GPU, we would still schedule
a record event on the input stream. That would delay the scheduling of the kernel
after all asyncMemcpy that had been scheduled before complete their execution,
reducing parallelism between execution and I/O.

With this change, the aim is to entirely skip recording the event on the input
stream and directly schedule the ready-to-execute kernel on an exec stream of
the GPU.

Note that the behavior of data already in transfer is unchanged: no additional
transfer is scheduled, but the task needs to wait on the input stream progress
to be ready to execute.